flask_backend/requirements.txt
acres==0.5.0
addict==2.4.0
annotated-types==0.7.0
attrs==25.3.0
babel==2.17.0
blinker==1.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
configobj==5.0.9
configparser==7.2.0
csvw==3.5.1
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.11
dlinfo==2.0.0
docopt==0.6.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
espeakng-loader==0.2.4
etelemetry==0.3.1
filelock==3.18.0
fitz==0.0.1.dev2
Flask==3.1.1
Flask-Cors
python-dotenv
gunicorn
fsspec==2025.5.1
httplib2==0.22.0
huggingface-hub==0.32.4
idna==3.10
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kokoro==0.9.4
langcodes==3.5.0
language-tags==1.2.0
language_data==1.3.0
loguru==0.7.3
looseversion==1.3.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
misaki==0.9.4
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
num2words==0.5.14
numpy==2.2.6
packaging==25.0
pandas==2.3.0
pathlib==1.0.1
phonemizer-fork==3.3.2
pillow==11.0.0
preshed==3.0.10
prov==2.0.1
puremagic==1.29
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydot==4.0.0
Pygments==2.19.1
PyMuPDF==1.26.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rich==14.0.0
rpds-py==0.25.1
safetensors==0.5.3
scipy==1.15.3
segments==2.3.0
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smart-open==7.1.0
soundfile==0.13.1
spacy==3.8.7
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.14.0
thinc==8.3.6
tokenizers==0.21.1
tqdm==4.67.1
traits==7.0.2
transformers==4.52.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
win32_setctime==1.2.0
wrapt==1.17.2
